adjust the models/folder structure for future extension
