+++ src/dsr/models/__init__.py
@@ -1,6 +1,7 @@
-"""Model architectures."""
+"""Model architectures and shared building blocks."""
 
+from . import components
 from .sedd import SEDD, SEDDConfig
 
 
-__all__ = ["SEDD", "SEDDConfig"]
+__all__ = ["SEDD", "SEDDConfig", "components"]
