+++ src/dsr/models/components/__init__.py
@@ -1,0 +1,19 @@
+"""Reusable building blocks shared across model architectures."""
+
+from .fused_ops import (
+    bias_dropout_add_scale_fused_inference,
+    bias_dropout_add_scale_fused_train,
+    modulate_fused,
+)
+from .mlp import MLP
+from .rotary import Rotary, apply_rotary_pos_emb
+
+
+__all__ = [
+    "bias_dropout_add_scale_fused_inference",
+    "bias_dropout_add_scale_fused_train",
+    "modulate_fused",
+    "MLP",
+    "Rotary",
+    "apply_rotary_pos_emb",
+]
