+++ src/dsr/models/sedd/__init__.py
@@ -1,0 +1,7 @@
+"""SEDD model package."""
+
+from .config import SEDDConfig
+from .model import SEDD
+
+
+__all__ = ["SEDD", "SEDDConfig"]
