+++ src/dsr/models/training/__init__.py
@@ -1,0 +1,6 @@
+"""Utilities that support model training workflows."""
+
+from .ema import ExponentialMovingAverage
+
+
+__all__ = ["ExponentialMovingAverage"]
